Move static scaffold app into templates
